MQTT: enable log capturing
